refactor(field-trip): fix positioning of field trip windows
