+++ lib/components/admin/call-taker-windows.js
@@ -14,38 +14,31 @@ class CallTakerWindows extends Component {
   render () {
     const {callTaker, fetchQueries, searches, toggleCallHistory} = this.props
     const {activeCall, callHistory} = callTaker
+    if (!callHistory.visible) return null
     return (
-      <>
-        {callHistory.visible
-          // Active call window
-          ? <DraggableWindow
-            draggableProps={{
-              defaultPosition: callHistory.position
-            }}
-            header={<h4><Icon type='history' /> Call history</h4>}
-            onClickClose={toggleCallHistory}
-          >
-            {activeCall
-              ? <CallRecord
-                call={activeCall}
-                searches={searches}
-                inProgress />
-              : null
-            }
-            {callHistory.calls.data.length > 0
-              ? callHistory.calls.data.map((call, i) => (
-                <CallRecord
-                  key={i}
-                  index={i}
-                  call={call}
-                  fetchQueries={fetchQueries} />
-              ))
-              : <div>No calls in history</div>
-            }
-          </DraggableWindow>
+      <DraggableWindow
+        header={<h4><Icon type='history' /> Call history</h4>}
+        onClickClose={toggleCallHistory}
+        style={{right: '15px', top: '50px'}}
+      >
+        {activeCall
+          ? <CallRecord
+            call={activeCall}
+            searches={searches}
+            inProgress />
           : null
         }
-      </>
+        {callHistory.calls.data.length > 0
+          ? callHistory.calls.data.map((call, i) => (
+            <CallRecord
+              key={i}
+              index={i}
+              call={call}
+              fetchQueries={fetchQueries} />
+          ))
+          : <div>No calls in history</div>
+        }
+      </DraggableWindow>
     )
   }
 }
