+++ lib/components/admin/draggable-window.js
@@ -13,6 +13,7 @@ export default class DraggableWindow extends Component {
       footer,
       header,
       height = '245px',
+      onClickClose,
       style
     } = this.props
     const GREY_BORDER = '#777 1.3px solid'
@@ -44,7 +45,7 @@ export default class DraggableWindow extends Component {
           >
             <button
               className='clear-button-formatting pull-right'
-              onClick={this.props.onClickClose}
+              onClick={onClickClose}
               style={{margin: '4px 4px 0 0'}}
             >
               <Icon type='times' />
