+++ lib/components/admin/field-trip-details.js
@@ -51,12 +51,12 @@ class FieldTripDetails extends Component {
   render () {
     const {
       addFieldTripNote,
-      callTaker,
       dateFormat,
       deleteFieldTripNote,
       request,
       setRequestGroupSize,
-      setRequestPaymentInfo
+      setRequestPaymentInfo,
+      style
     } = this.props
     if (!request) return null
     const {
@@ -69,20 +69,15 @@ class FieldTripDetails extends Component {
       ticketType,
       travelDate
     } = request
-    const {fieldTrip} = callTaker
-    const defaultPosition = {...fieldTrip.position}
     const internalNotes = []
     const operationalNotes = []
     notes && notes.forEach(note => {
       if (note.type === 'internal') internalNotes.push(note)
       else operationalNotes.push(note)
     })
-    defaultPosition.x = defaultPosition.x - 460
-    defaultPosition.y = defaultPosition.y - 100
     const travelDateAsMoment = moment(travelDate)
     return (
       <DraggableWindow
-        draggableProps={{ defaultPosition }}
         footer={
           <div style={{padding: '5px 10px 0px 10px'}}>
             <DropdownButton
@@ -128,7 +123,7 @@ class FieldTripDetails extends Component {
         }
         height='375px'
         onClickClose={this._onCloseActiveFieldTrip}
-        style={{width: '450px'}}
+        style={style}
       >
         <Container>
           <Header>Group Information</Header>
@@ -194,7 +189,6 @@ const mapStateToProps = (state, ownProps) => {
   const {activeId, requests} = state.callTaker.fieldTrip
   const request = requests.data.find(req => req.id === activeId)
   return {
-    callTaker: state.callTaker,
     currentQuery: state.otp.currentQuery,
     datastoreUrl: state.otp.config.datastoreUrl,
     dateFormat: getDateFormat(state.otp.config),
