+++ lib/reducers/call-taker.js
@@ -1,25 +1,18 @@
-import coreUtils from '@opentripplanner/core-utils'
 import update from 'immutability-helper'
 import moment from 'moment'
 
+import {constructNewCall} from '../util/call-taker'
 import {FETCH_STATUS} from '../util/constants'
-import {getTimestamp} from '../util/state'
-
-const { randId } = coreUtils.storage
-
-const UPPER_RIGHT_CORNER = {x: 604, y: 53}
-const LOWER_RIGHT_CORNER = {x: 504, y: 365}
 
 function createCallTakerReducer () {
   const initialState = {
     activeCall: null,
     callHistory: {
-      position: UPPER_RIGHT_CORNER,
-      visible: false,
       calls: {
         status: FETCH_STATUS.UNFETCHED,
         data: []
-      }
+      },
+      visible: false
     },
     fieldTrip: {
       activeId: null,
@@ -27,7 +20,6 @@ function createCallTakerReducer () {
         tab: 'new'
       },
       groupSize: null,
-      position: LOWER_RIGHT_CORNER,
       requests: {
         status: FETCH_STATUS.UNFETCHED,
         data: []
@@ -39,11 +31,7 @@ function createCallTakerReducer () {
   return (state = initialState, action) => {
     switch (action.type) {
       case 'BEGIN_CALL': {
-        const newCall = {
-          startTime: getTimestamp(),
-          id: randId(),
-          searches: []
-        }
+        const newCall = constructNewCall()
         // Initialize new call and show call history window.
         return update(state, {
           activeCall: { $set: newCall },
@@ -115,13 +103,20 @@ function createCallTakerReducer () {
         })
       }
       case 'ROUTING_RESPONSE': {
-        // If call is in progress, record search ID when a routing response is
-        // fulfilled.
-        // TODO: How should we handle routing errors.
         if (state.activeCall) {
+          // If call is in progress, record search ID when a routing response is
+          // fulfilled.
+          // TODO: How should we handle routing errors.
           return update(state, {
             activeCall: { searches: { $push: [action.payload.searchId] } }
           })
+        } else if (state.callHistory.visible) {
+          // If call not in progress, but history is visible,
+          // construct new call and add search.
+          const newCall = constructNewCall()
+          newCall.searches.push(action.payload.searchId)
+          // Initialize new call and show call history window.
+          return update(state, { activeCall: { $set: newCall } })
         }
         // Otherwise, ignore.
         return state
@@ -144,7 +139,10 @@ function createCallTakerReducer () {
       }
       case 'TOGGLE_FIELD_TRIPS': {
         return update(state, {
-          fieldTrip: { visible: { $set: !state.fieldTrip.visible } }
+          fieldTrip: {
+            activeId: { $set: null },
+            visible: { $set: !state.fieldTrip.visible }
+          }
         })
       }
       case 'END_CALL': {
