+++ lib/util/call-taker.js
@@ -1,6 +1,8 @@
 import {isTransit} from '@opentripplanner/core-utils/lib/itinerary'
+import {randId} from '@opentripplanner/core-utils/lib/storage'
 
 import {getRoutingParams} from '../actions/api'
+import {getTimestamp} from './state'
 
 export const TICKET_TYPES = {
   own_tickets: 'Will use own tickets',
@@ -35,6 +37,14 @@ export const PAYMENT_FIELDS = [
   {label: 'Credit card last 4 digits', fieldName: 'ccLastFour'},
   {label: 'Check/Money order number', fieldName: 'checkNumber'}
 ]
+
+export function constructNewCall () {
+  return {
+    startTime: getTimestamp(),
+    id: randId(),
+    searches: []
+  }
+}
 
 function placeToLatLonStr (place) {
   return `${place.lat.toFixed(6)},${place.lon.toFixed(6)}`
@@ -105,7 +115,7 @@ export function createTripPlan (planData, queryParams) {
     itineraries: []
   }
   if (!planData) return
-  tripPlan.itineraries = tripPlan.planData.map(itinData => 
+  tripPlan.itineraries = tripPlan.planData.map(itinData =>
     createItinerary(itinData, tripPlan))
   const timeBounds = calculateTimeBounds(tripPlan.itineraries)
   return {...tripPlan, ...timeBounds}
